Complete migration.
