+++ qualtran/bloqs/chemistry/trotter/__init__.py
@@ -11,5 +11,7 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
+"""Bloqs for Trotter simulation of the real space grid Hamiltonian."""
 
 from .kinetic import KineticEnergy
+from .potential import PairPotential, PotentialEnergy
