+++ qualtran/bloqs/chemistry/trotter/qvr_test.py
@@ -11,7 +11,7 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-from qualtran.bloqs.chemistry.trotter import QuantumVariableRotation
+from qualtran.bloqs.chemistry.trotter.qvr import QuantumVariableRotation
 
 
 def test_qvr():
